+++ komapper-core/src/main/kotlin/org/komapper/core/dsl/query/EntityUpsertQueryBuilder.kt
@@ -13,24 +13,24 @@ import org.komapper.core.dsl.metamodel.EntityMetamodel
  */
 @ThreadSafe
 internal interface EntityUpsertQueryBuilder<ENTITY : Any, ID : Any, META : EntityMetamodel<ENTITY, ID, META>> {
-    fun single(entity: ENTITY): EntityUpsertQuery<Int>
-    fun multiple(entities: List<ENTITY>): EntityUpsertQuery<Int>
-    fun batch(entities: List<ENTITY>, batchSize: Int? = null): EntityUpsertQuery<List<Int>>
+    fun single(entity: ENTITY): EntityUpsertQuery<Long>
+    fun multiple(entities: List<ENTITY>): EntityUpsertQuery<Long>
+    fun batch(entities: List<ENTITY>, batchSize: Int? = null): EntityUpsertQuery<List<Long>>
 }
 
 internal data class EntityUpsertQueryBuilderImpl<ENTITY : Any, ID : Any, META : EntityMetamodel<ENTITY, ID, META>>(
     private val context: EntityUpsertContext<ENTITY, ID, META>,
 ) : EntityUpsertQueryBuilder<ENTITY, ID, META> {
 
-    override fun single(entity: ENTITY): EntityUpsertQuery<Int> {
+    override fun single(entity: ENTITY): EntityUpsertQuery<Long> {
         return EntityUpsertSingleQuery(context, entity)
     }
 
-    override fun multiple(entities: List<ENTITY>): EntityUpsertQuery<Int> {
+    override fun multiple(entities: List<ENTITY>): EntityUpsertQuery<Long> {
         return EntityUpsertMultipleQuery(context, entities)
     }
 
-    override fun batch(entities: List<ENTITY>, batchSize: Int?): EntityUpsertQuery<List<Int>> {
+    override fun batch(entities: List<ENTITY>, batchSize: Int?): EntityUpsertQuery<List<Long>> {
         val context = if (batchSize != null) {
             val options = context.insertContext.options.copy(batchSize = batchSize)
             val insertContext = context.insertContext.copy(options = options)
