+++ komapper-core/src/main/kotlin/org/komapper/core/dsl/query/RelationInsertQuery.kt
@@ -24,7 +24,7 @@ interface RelationInsertQuery<ENTITY : Any, ID : Any, META : EntityMetamodel<ENT
      * @param declaration the assignment declaration
      * @return the query
      */
-    fun values(declaration: AssignmentDeclaration<ENTITY, META>): RelationInsertQuery<ENTITY, ID, META, Pair<Int, ID?>>
+    fun values(declaration: AssignmentDeclaration<ENTITY, META>): RelationInsertQuery<ENTITY, ID, META, Pair<Long, ID?>>
 
     /**
      * Builds a query with the options applied.
@@ -37,14 +37,14 @@ interface RelationInsertQuery<ENTITY : Any, ID : Any, META : EntityMetamodel<ENT
 
 internal data class RelationInsertSelectQuery<ENTITY : Any, ID : Any, META : EntityMetamodel<ENTITY, ID, META>>(
     private val context: RelationInsertSelectContext<ENTITY, ID, META>,
-) : RelationInsertQuery<ENTITY, ID, META, Pair<Int, List<ID>>> {
+) : RelationInsertQuery<ENTITY, ID, META, Pair<Long, List<ID>>> {
 
-    override fun options(configure: (InsertOptions) -> InsertOptions): RelationInsertQuery<ENTITY, ID, META, Pair<Int, List<ID>>> {
+    override fun options(configure: (InsertOptions) -> InsertOptions): RelationInsertQuery<ENTITY, ID, META, Pair<Long, List<ID>>> {
         val newContext = context.copy(options = configure(context.options))
         return copy(context = newContext)
     }
 
-    override fun values(declaration: AssignmentDeclaration<ENTITY, META>): RelationInsertQuery<ENTITY, ID, META, Pair<Int, ID?>> {
+    override fun values(declaration: AssignmentDeclaration<ENTITY, META>): RelationInsertQuery<ENTITY, ID, META, Pair<Long, ID?>> {
         val newContext = context.asRelationInsertValuesContext(declaration)
         return RelationInsertValuesQuery(newContext)
     }
@@ -56,14 +56,14 @@ internal data class RelationInsertSelectQuery<ENTITY : Any, ID : Any, META : Ent
 
 internal data class RelationInsertValuesQuery<ENTITY : Any, ID : Any, META : EntityMetamodel<ENTITY, ID, META>>(
     private val context: RelationInsertValuesContext<ENTITY, ID, META>,
-) : RelationInsertQuery<ENTITY, ID, META, Pair<Int, ID?>> {
+) : RelationInsertQuery<ENTITY, ID, META, Pair<Long, ID?>> {
 
-    override fun options(configure: (InsertOptions) -> InsertOptions): RelationInsertQuery<ENTITY, ID, META, Pair<Int, ID?>> {
+    override fun options(configure: (InsertOptions) -> InsertOptions): RelationInsertQuery<ENTITY, ID, META, Pair<Long, ID?>> {
         val newContext = context.copy(options = configure(context.options))
         return copy(context = newContext)
     }
 
-    override fun values(declaration: AssignmentDeclaration<ENTITY, META>): RelationInsertQuery<ENTITY, ID, META, Pair<Int, ID?>> {
+    override fun values(declaration: AssignmentDeclaration<ENTITY, META>): RelationInsertQuery<ENTITY, ID, META, Pair<Long, ID?>> {
         val newContext = context.copy(values = context.values + declaration)
         return copy(context = newContext)
     }
