+++ komapper-jdbc/src/main/kotlin/org/komapper/jdbc/dsl/runner/JdbcEntityUpsertSingleRunner.kt
@@ -10,7 +10,7 @@ import org.komapper.jdbc.JdbcDatabaseConfig
 internal class JdbcEntityUpsertSingleRunner<ENTITY : Any, ID : Any, META : EntityMetamodel<ENTITY, ID, META>>(
     context: EntityUpsertContext<ENTITY, ID, META>,
     private val entity: ENTITY,
-) : JdbcRunner<Int> {
+) : JdbcRunner<Long> {
 
     private val runner: EntityUpsertSingleRunner<ENTITY, ID, META> =
         EntityUpsertSingleRunner(context, entity)
@@ -22,7 +22,7 @@ internal class JdbcEntityUpsertSingleRunner<ENTITY : Any, ID : Any, META : Entit
         runner.check(config)
     }
 
-    override fun run(config: JdbcDatabaseConfig): Int {
+    override fun run(config: JdbcDatabaseConfig): Long {
         val newEntity = preUpsert(config, entity)
         val (count) = upsert(config, newEntity)
         return count
@@ -32,7 +32,7 @@ internal class JdbcEntityUpsertSingleRunner<ENTITY : Any, ID : Any, META : Entit
         return support.preUpsert(config, entity)
     }
 
-    private fun upsert(config: JdbcDatabaseConfig, entity: ENTITY): Pair<Int, List<Long>> {
+    private fun upsert(config: JdbcDatabaseConfig, entity: ENTITY): Pair<Long, List<Long>> {
         val statement = runner.buildStatement(config, entity)
         return support.upsert(config, false) { it.executeUpdate(statement) }
     }
